fixed Travis remarks

+++ exchanges/coinbasepro/coinbasepro_wrapper.go
@@ -619,11 +619,10 @@ func (c *CoinbasePro) AuthenticateWebsocket() error {
 
 // GetHistoricCandles Allows to retrieve an amount of candles back in time starting from now up to rangesize * granularity, where granularity is the trade period covered by each candle
 func (c *CoinbasePro) GetHistoricCandles(p currency.Pair, rangesize, granularity int64) ([]exchange.Candle, error) {
-	//start end granularity
 	end := time.Now().UTC()
 	b := granularity * rangesize
 	start := time.Now().UTC().Add(-time.Second * time.Duration(b))
-	history, err := c.GetHistoricRates(p.String(), start.Format(time.RFC3339), end.Format(time.RFC3339), int64(granularity))
+	history, err := c.GetHistoricRates(p.String(), start.Format(time.RFC3339), end.Format(time.RFC3339), granularity)
 	if err != nil {
 		return nil, err
 	}
